Replace init.d script for inetd by Debian version
This script still has problems stopping the service, though. But it's
just an example, anyway, and the important functions for preprocessing
inetd.conf already work.

BTW, they had to be changed to work under Debian, too. Only IPv4 is
supported for now, because it was necessary to parse the IP address more
directly.

There are just too many variants of "ifconfig" out there which insert
all kinds of fixed text between "inet" and the IP address in their
output.

And the only thing common to interfaces in that output seems to be that
their names always start in the first column.